+++ invoices/tools/src/export_invoices.js
@@ -76,6 +76,7 @@ function generateCsvInvoices(invoicesTable) {
                 } 
                 
                 for (let j = 0; j < invoiceObj.items.length; j++) {
+                    let invoiceDiscount = "";
                     let itemTotal = "";
                     let itemUnitPrice = "";
                     let itemDiscount = "";
@@ -85,6 +86,7 @@ function generateCsvInvoices(invoicesTable) {
                         return "";
                     }
                     if (invoiceObj.document_info.vat_mode === "vat_excl") {
+                        invoiceDiscount = invoiceObj.billing_info.total_discount_vat_exclusive;
                         if (!invoiceObj.items[j].total_amount_vat_exclusive) {
                             row.addMessage(qsTr("%1 is a required field").arg("ItemTotal"), "RowId", "missing_field");
                             rowMatched = false;
@@ -95,6 +97,7 @@ function generateCsvInvoices(invoicesTable) {
                         }
                         
                     } else {
+                        invoiceDiscount = invoiceObj.billing_info.total_discount_vat_inclusive;
                         if (!invoiceObj.items[j].total_amount_vat_inclusive) {
                             row.addMessage(qsTr("%1 is a required field").arg("ItemTotal"), "RowId", "missing_field");
                             rowMatched = false;
@@ -119,7 +122,7 @@ function generateCsvInvoices(invoicesTable) {
                         itemDescription = `"${itemDescription}"`;
                     }
                         
-                    csv += `${getValue(invoiceObj.document_info.number)},${getValue(invoiceObj.document_info.date)},${getValue(invoiceObj.payment_info.due_date)},${getValue(invoiceObj.document_info.description)},${getValue(invoiceObj.billing_info.total_discount_vat_inclusive)},${getValue(invoiceObj.billing_info.total_vat_amount)},${getValue(invoiceObj.billing_info.total_to_pay)},${getValue(invoiceObj.document_info.currency)},${getValue(invoiceObj.document_info.rounding_total)},${getValue(invoiceObj.document_info.vat_mode)},`+
+                    csv += `${getValue(invoiceObj.document_info.number)},${getValue(invoiceObj.document_info.date)},${getValue(invoiceObj.payment_info.due_date)},${getValue(invoiceObj.document_info.description)},${getValue(invoiceDiscount)},${getValue(invoiceObj.billing_info.total_vat_amount)},${getValue(invoiceObj.billing_info.total_to_pay)},${getValue(invoiceObj.document_info.currency)},${getValue(invoiceObj.document_info.rounding_total)},${getValue(invoiceObj.document_info.vat_mode)},`+
                            `${getValue(invoiceObj.customer_info.number)},${getValue(invoiceObj.customer_info.first_name)} ${getValue(invoiceObj.customer_info.last_name)},${getValue(invoiceObj.items[j].number)},${getValue(itemDescription)},${getValue(invoiceObj.items[j].quantity)},${itemUnitPrice},${getValue(invoiceObj.items[j].mesure_unit)},${getValue(invoiceObj.items[j].unit_price.vat_rate)},${getValue(invoiceObj.items[j].unit_price.vat_code)},${getValue(itemDiscount)},${getValue(itemTotal)},${getValue(invoiceObj.items[j].total_vat_amount)}\n`;
                 }
             }